pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autosar-calltree"
version = "0.2.0"
description = "A Python tool to analyze C/AUTOSAR codebases and generate function call trees with Mermaid and XMI output"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "melodypapa", email = "melodypapa@outlook.com"}
]
maintainers = [
    {name = "melodypapa", email = "melodypapa@outlook.com"}
]
keywords = [
    "autosar",
    "c",
    "call-tree",
    "static-analysis",
    "code-analysis",
    "mermaid",
    "xmi",
    "uml",
    "embedded",
    "automotive"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Documentation",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "isort>=5.10.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-click>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/autosar-calltree"
Documentation = "https://autosar-calltree.readthedocs.io"
Repository = "https://github.com/yourusername/autosar-calltree"
"Bug Tracker" = "https://github.com/yourusername/autosar-calltree/issues"

[project.scripts]
calltree = "autosar_calltree.cli:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
autosar_calltree = ["templates/*"]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=autosar_calltree --cov-report=html --cov-report=term"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = 3.9
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
